Replaced confirmation prompt

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageParams } from '../_models/paginationParams';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   messageParams: MessageParams;
   loadingFlag = false;
 
-  constructor(private messageService: MessageService) {
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) {
     this.messageParams = this.messageService.getMessageParams();
   }
 
@@ -42,7 +43,11 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() =>
-      this.messages.splice(this.messages.findIndex(m => m.id == id), 1));
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(r => {
+      if (r) {
+        this.messageService.deleteMessage(id).subscribe(() =>
+          this.messages.splice(this.messages.findIndex(m => m.id == id), 1));
+      }
+    });
   }
 }
